+++ packages/insomnia/src/ui/components/request-url-bar.tsx
@@ -18,8 +18,8 @@ import * as network from '../../network/network';
 import { convert } from '../../utils/importers/convert';
 import { buildQueryStringFromParams, joinUrlAndQueryString } from '../../utils/url/querystring';
 import { SegmentEvent } from '../analytics';
-import { useRequestMetaPatcher } from '../hooks/create-request';
 import { useReadyState } from '../hooks/use-ready-state';
+import { useRequestMetaPatcher } from '../hooks/use-request';
 import { useTimeoutWhen } from '../hooks/useTimeoutWhen';
 import { RequestLoaderData } from '../routes/request';
 import { RootLoaderData } from '../routes/root';
